+++ src/test/scala/mesosphere/mesos/PortWithRoleRandomPortsFromRangesTest.scala
@@ -1,9 +1,9 @@
-package mesosphere.marathon.tasks
+package mesosphere.mesos
 
-import java.util
+import java.{ util => jutil }
 import java.util.concurrent.TimeUnit
 
-import mesosphere.marathon.tasks.PortsMatcher.{ PortRange }
+import mesosphere.mesos.PortsMatcher.PortRange
 import mesosphere.mesos.PortsMatchResult.PortWithRole
 import mesosphere.marathon.test.MarathonSpec
 import org.slf4j.LoggerFactory
@@ -23,7 +23,7 @@ class PortWithRoleRandomPortsFromRangesTest extends MarathonSpec {
   private[this] def withRandomSeeds(input: Seq[PortRange], expectedOutput: Iterable[PortWithRole]): Unit = {
     for (seed <- 1 to 10) {
       withClue(s"seed = $seed") {
-        val rand = new Random(new util.Random(seed.toLong))
+        val rand = new Random(new jutil.Random(seed.toLong))
 
         assert(PortsMatcher.lazyRandomPortsFromRanges(rand)(input).to[Set] == expectedOutput.to[Set])
       }
@@ -103,7 +103,7 @@ class PortWithRoleRandomPortsFromRangesTest extends MarathonSpec {
     val numberOfRanges = 500
 
     def performTest(seed: Int): Unit = {
-      val rand = new Random(new util.Random(seed.toLong))
+      val rand = new Random(new jutil.Random(seed.toLong))
       val ports = PortsMatcher.lazyRandomPortsFromRanges(rand)(
         (0 to numberOfRanges).map { i =>
           val rangeSize: Long = Integer.MAX_VALUE.toLong / numberOfRanges.toLong
